+++ test/integration/instance.validations.test.js
@@ -131,6 +131,45 @@ describe(Support.getTestDialectTeaser('InstanceValidator'), function() {
           expect(err.errors[0].message).to.equal('custom unique error message');
         });
     });
+
+    it('should handle multiple unique messages correctly', function() {
+      var Model = this.sequelize.define('model', {
+        uniqueName1: {
+          type: Sequelize.STRING,
+          unique: { msg: 'custom unique error message 1' }
+        },
+        uniqueName2: {
+          type: Sequelize.STRING,
+          unique: { msg: 'custom unique error message 2' }
+        },
+      });
+      var records = [
+        { uniqueName1: 'unique name one', uniqueName2: 'unique name one' },
+        { uniqueName1: 'unique name one', uniqueName2: 'this is ok' },
+        { uniqueName1: 'this is ok', uniqueName2: 'unique name one' },
+      ];
+      return Model.sync({ force: true })
+        .then(function() {
+          return Model.create(records[0]);
+        }).then(function(instance) {
+          expect(instance).to.be.ok;
+          return expect(Model.create(records[1])).to.be.rejected;
+        }).then(function(err) {
+          expect(err).to.be.an.instanceOf(Error);
+          console.log(err.errors);
+          expect(err.errors).to.have.length(1);
+          expect(err.errors[0].path).to.include('uniqueName1');
+          expect(err.errors[0].message).to.equal('custom unique error message 1');
+
+          return expect(Model.create(records[2])).to.be.rejected;
+        }).then(function(err) {
+          expect(err).to.be.an.instanceOf(Error);
+          console.log(err.errors);
+          expect(err.errors).to.have.length(1);
+          expect(err.errors[0].path).to.include('uniqueName2');
+          expect(err.errors[0].message).to.equal('custom unique error message 2');
+        });
+    });
   });
 
   describe('#create', function() {